Redid the sprite system
Now the sprites are divided in to groups (currently ItemsSprites,
TileSprites and DefaultSprites). To have it more organized. To access a
sprite you call ResourceManager.GetSprite(std::string SpriteType,
std::string/int ID).

TODO:

* Add a debug flag you can send to the textbuffer function that tiles
the debug text on eachother so I don't have to manually change the x and
y coordinates of the text.